Add openDebug command to the setup API
